Move inlink count tables to tools database
The old database servers on Toolforge (c1 and c3) are being decomissioned.
New servers are online, but do not allow creation of databases, meaning our
inlink table needed to move to the Tools database server. This commit
defines the tables on the Tools DB server, and updates the script that
populates those tables so data is grabbed from the new analytics-
oriented replicated Wikipedia database servers.