Mark redacted edits as reverted/unhelpful
If an edit comment has been redacted, the edit is now marked as
a "reverted" edit, and will be discarded if the bot is set up
to do so. This seems sensible as a deleted edit comment suggests
the edit was unproductive.

This fixes a bug where a user with any deleted edit comment
would result in a KeyError exception as those edits were not
labelled either "keep" or "revert" and thus would not be found
when trying to discard reverted edits.