Update opentasks to support Python 3
This updates opentasks.py so it is Python 3 compliant, and sets up
a shell script to launch it with a Python 3.5 virtual environment
specific to this part of SuggestBot.

We also add YAML as the configuration file format and move all
settings into that file. Configuration files for the four wikis
we're supporting (English, Polish, Russian, and Thai) are added.
The task specification format is now properly documented in the
configuration files.

Support for relisted AfDs are removed as they're no longer used.